Moved string extensions, added test folders
